+++ Assets/MotionFramework/ThirdParty/YooAsset/Editor/AssetBundleBuilder/AssetBundleBuilderWindow.cs
@@ -29,8 +29,7 @@ namespace YooAsset.Editor
 		private PopupField<string> _encryptionField;
 		private EnumField _compressionField;
 		private Toggle _appendExtensionToggle;
-		private Toggle _copyBuildinTagFilesToggle;
-		
+
 		public void CreateGUI()
 		{
 			try
@@ -41,7 +40,7 @@ namespace YooAsset.Editor
 				var visualAsset = EditorHelper.LoadWindowUXML<AssetBundleBuilderWindow>();
 				if (visualAsset == null)
 					return;
-				
+
 				visualAsset.CloneTree(root);
 
 				_buildTarget = EditorUserBuildSettings.activeBuildTarget;
@@ -108,7 +107,7 @@ namespace YooAsset.Editor
 				_compressionField = root.Q<EnumField>("Compression");
 				_compressionField.Init(AssetBundleBuilderSettingData.Setting.CompressOption);
 				_compressionField.SetValueWithoutNotify(AssetBundleBuilderSettingData.Setting.CompressOption);
-				_compressionField.style.width = 300;	
+				_compressionField.style.width = 300;
 				_compressionField.RegisterValueChangedCallback(evt =>
 				{
 					AssetBundleBuilderSettingData.Setting.CompressOption = (ECompressOption)_compressionField.value;
@@ -120,14 +119,6 @@ namespace YooAsset.Editor
 				_appendExtensionToggle.RegisterValueChangedCallback(evt =>
 				{
 					AssetBundleBuilderSettingData.Setting.AppendExtension = _appendExtensionToggle.value;
-				});
-
-				// 拷贝首包文件
-				_copyBuildinTagFilesToggle = root.Q<Toggle>("CopyBuildinFiles");
-				_copyBuildinTagFilesToggle.SetValueWithoutNotify(AssetBundleBuilderSettingData.Setting.CopyBuildinTagFiles);
-				_copyBuildinTagFilesToggle.RegisterValueChangedCallback(evt =>
-				{
-					AssetBundleBuilderSettingData.Setting.CopyBuildinTagFiles = _copyBuildinTagFilesToggle.value;
 				});
 
 				// 构建按钮
@@ -154,7 +145,6 @@ namespace YooAsset.Editor
 			_encryptionField.SetEnabled(enableElement);
 			_compressionField.SetEnabled(enableElement);
 			_appendExtensionToggle.SetEnabled(enableElement);
-			_copyBuildinTagFilesToggle.SetEnabled(buildMode == EBuildMode.ForceRebuild || buildMode == EBuildMode.IncrementalBuild);
 		}
 		private void BuildButton_clicked()
 		{
@@ -175,22 +165,28 @@ namespace YooAsset.Editor
 		/// </summary>
 		private void ExecuteBuild()
 		{
+			var buildMode = (EBuildMode)_buildModeField.value;
+
 			string defaultOutputRoot = AssetBundleBuilderHelper.GetDefaultOutputRoot();
 			BuildParameters buildParameters = new BuildParameters();
 			buildParameters.OutputRoot = defaultOutputRoot;
 			buildParameters.BuildTarget = _buildTarget;
-			buildParameters.BuildMode = (EBuildMode)_buildModeField.value;
+			buildParameters.BuildMode = buildMode;
 			buildParameters.BuildVersion = _buildVersionField.value;
 			buildParameters.BuildinTags = _buildinTagsField.value;
 			buildParameters.VerifyBuildingResult = true;
 			buildParameters.EnableAddressable = AssetBundleCollectorSettingData.Setting.EnableAddressable;
 			buildParameters.AppendFileExtension = _appendExtensionToggle.value;
-			buildParameters.CopyBuildinTagFiles = _copyBuildinTagFilesToggle.value;
+			buildParameters.CopyBuildinTagFiles = buildMode == EBuildMode.ForceRebuild;
 			buildParameters.EncryptionServices = CreateEncryptionServicesInstance();
 			buildParameters.CompressOption = (ECompressOption)_compressionField.value;
 
 			AssetBundleBuilder builder = new AssetBundleBuilder();
-			builder.Run(buildParameters);
+			bool succeed = builder.Run(buildParameters);
+			if (succeed)
+			{
+				EditorUtility.RevealInFinder($"{buildParameters.OutputRoot}/{buildParameters.BuildTarget}/{buildParameters.BuildVersion}");
+			}
 		}
 
 		// 加密类相关