+++ Assets/MotionFramework/ThirdParty/YooAsset/CHANGELOG.md
@@ -2,6 +2,109 @@
 
 All notable changes to this package will be documented in this file.
 
+## [1.1.1] - 2022-07-07
+
+### Fixed
+
+- 修复了AssetBundleDebugger窗口，View下拉页签切换无效的问题。
+- 修复了在Unity2020.3版本下UniTask在真机上的一个IL2CPP相关的错误。
+
+### Changed
+
+- 优化了AssetBundleDebugger窗口，增加了帧数显示以及回放功能。
+- 优化了AssetBundleBuilder的代码结构。
+- 增强了YooAssets.GetRawFileAsync()方法的容错。
+
+### Added
+
+- 新增了OperationHandleBase.GetAssetInfo()方法。
+
+  ````c#
+  /// <summary>
+  /// 获取资源信息
+  /// </summary>
+  public AssetInfo GetAssetInfo();
+  ````
+
+- 新增了AssetOperationHandle.GetAssetObjet<TAsset>()方法。
+
+  ````c#
+  /// <summary>
+  /// 获取资源对象
+  /// </summary>
+  /// <typeparam name="TAsset">资源类型</typeparam>
+  public TAsset GetAssetObjet<TAsset>()；
+  ````
+
+- 新增了弱联网情况下加载补丁清单方法。
+
+  ````c#
+  /// <summary>
+  /// 弱联网情况下加载补丁清单
+  /// 注意：当指定版本内容验证失败后会返回失败。
+  /// </summary>
+  /// <param name="resourceVersion">指定的资源版本</param>
+  public static UpdateManifestOperation WeaklyUpdateManifestAsync(int resourceVersion)；
+  ````
+
+### Removed
+
+- 离线运行模式（OfflinePlayMode）下移除了资内置资源解压相关逻辑。
+- 移除了初始化参数：AutoReleaseGameObjectHandle及相关代码逻辑。
+
+## [1.1.0] - 2022-06-23
+
+### Fixed
+
+- 修复了AssetBundleCollector窗口，在切换EnableAddressable时未及时刷新界面的问题。
+- 修复了AssetBundleCollector窗口，资源过滤器CollectSprite无效的问题。
+- 修复了AssetBundleCollector窗口，无法正常预览StaticAssetCollector的资源列表的问题。
+- 修复了在离线模式下原生文件每次都从包内加载的问题。
+
+### Changed
+
+- 变更了共享资源打包机制。
+- AssetBundleCollector窗口增加了分组禁用功能。
+- AssetBundleDebugger窗口增加了真机远程调试功能。
+- AssetBundleBuilder窗口在构建成功后自动显示构建文件夹。
+- DownloaderOperation.OnDownloadFileFailedCallback委托变更为OnDownloadErrorCallback委托。
+
+### Added
+
+- 新增UpdateManifestOperation.FoundNewManifest字段。
+- 新增DownloaderOperation.OnStartDownloadFileCallback委托。
+- 新增AssetInfo.Address字段。
+- 新增YooAssets.IsInitialized字段。
+- 新增YooAssets初始化参数。
+
+  ````c#
+  /// <summary>
+  /// 下载文件校验等级
+  /// </summary>
+  public EVerifyLevel VerifyLevel = EVerifyLevel.High;
+  ````
+
+- 新增YooAssets获取资源完成路径的方法。
+
+  ````c#
+  /// <summary>
+  /// 获取资源路径
+  /// </summary>
+  /// <param name="location">资源的定位地址</param>
+  /// <returns>如果location地址无效，则返回空字符串</returns>
+  public static string GetAssetPath(string location);
+  ````
+
+- 新增YooAssets初始化参数。
+
+  ```c#
+  /// <summary>
+  /// 自动释放游戏对象所属资源句柄
+  /// 说明：通过资源句柄实例化的游戏对象在销毁之后，会自动释放所属资源句柄。
+  /// </summary>
+  public bool AutoReleaseGameObjectHandle = false;
+  ```
+
 ## [1.0.10] - 2022-05-22
 
 ### Fixed